separate pager and plain apis for exchange
Separates the pager and enumerattion functionality
from the rest of the exchange api funcs for each
category.  Purely code movement, no logic changes.
