+++ src/pkg/services/m365/api/events.go
@@ -18,7 +18,6 @@ import (
 	"github.com/alcionai/corso/src/internal/m365/graph"
 	"github.com/alcionai/corso/src/pkg/backup/details"
 	"github.com/alcionai/corso/src/pkg/fault"
-	"github.com/alcionai/corso/src/pkg/path"
 )
 
 // ---------------------------------------------------------------------------
@@ -191,86 +190,6 @@ func (c Events) PatchCalendar(
 }
 
 // ---------------------------------------------------------------------------
-// container pager
-// ---------------------------------------------------------------------------
-
-// EnumerateContainers iterates through all of the users current
-// calendars, converting each to a graph.CacheFolder, and
-// calling fn(cf) on each one.
-// Folder hierarchy is represented in its current state, and does
-// not contain historical data.
-func (c Events) EnumerateContainers(
-	ctx context.Context,
-	userID, baseContainerID string,
-	fn func(graph.CachedContainer) error,
-	errs *fault.Bus,
-) error {
-	var (
-		el     = errs.Local()
-		config = &users.ItemCalendarsRequestBuilderGetRequestConfiguration{
-			QueryParameters: &users.ItemCalendarsRequestBuilderGetQueryParameters{
-				Select: idAnd("name"),
-			},
-		}
-		builder = c.Stable.
-			Client().
-			Users().
-			ByUserId(userID).
-			Calendars()
-	)
-
-	for {
-		if el.Failure() != nil {
-			break
-		}
-
-		resp, err := builder.Get(ctx, config)
-		if err != nil {
-			return graph.Stack(ctx, err)
-		}
-
-		for _, cal := range resp.GetValue() {
-			if el.Failure() != nil {
-				break
-			}
-
-			cd := CalendarDisplayable{Calendarable: cal}
-			if err := graph.CheckIDAndName(cd); err != nil {
-				errs.AddRecoverable(ctx, graph.Stack(ctx, err).Label(fault.LabelForceNoBackupCreation))
-				continue
-			}
-
-			fctx := clues.Add(
-				ctx,
-				"container_id", ptr.Val(cal.GetId()),
-				"container_name", ptr.Val(cal.GetName()))
-
-			temp := graph.NewCacheFolder(
-				cd,
-				path.Builder{}.Append(ptr.Val(cd.GetId())),          // storage path
-				path.Builder{}.Append(ptr.Val(cd.GetDisplayName()))) // display location
-			if err := fn(&temp); err != nil {
-				errs.AddRecoverable(ctx, graph.Stack(fctx, err).Label(fault.LabelForceNoBackupCreation))
-				continue
-			}
-		}
-
-		link, ok := ptr.ValOK(resp.GetOdataNextLink())
-		if !ok {
-			break
-		}
-
-		builder = users.NewItemCalendarsRequestBuilder(link, c.Stable.Adapter())
-	}
-
-	return el.Failure()
-}
-
-const (
-	eventBetaDeltaURLTemplate = "https://graph.microsoft.com/beta/users/%s/calendars/%s/events/delta"
-)
-
-// ---------------------------------------------------------------------------
 // items
 // ---------------------------------------------------------------------------
 
@@ -432,154 +351,6 @@ func (c Events) PostLargeAttachment(
 	}
 
 	return us, nil
-}
-
-// ---------------------------------------------------------------------------
-// item pager
-// ---------------------------------------------------------------------------
-
-var _ itemPager = &eventPager{}
-
-type eventPager struct {
-	gs      graph.Servicer
-	builder *users.ItemCalendarsItemEventsRequestBuilder
-	options *users.ItemCalendarsItemEventsRequestBuilderGetRequestConfiguration
-}
-
-func (c Events) NewEventPager(
-	ctx context.Context,
-	userID, containerID string,
-	immutableIDs bool,
-) (itemPager, error) {
-	options := &users.ItemCalendarsItemEventsRequestBuilderGetRequestConfiguration{
-		Headers: newPreferHeaders(preferPageSize(maxNonDeltaPageSize), preferImmutableIDs(immutableIDs)),
-	}
-
-	builder := c.Stable.
-		Client().
-		Users().
-		ByUserId(userID).
-		Calendars().
-		ByCalendarId(containerID).
-		Events()
-
-	return &eventPager{c.Stable, builder, options}, nil
-}
-
-func (p *eventPager) getPage(ctx context.Context) (DeltaPageLinker, error) {
-	resp, err := p.builder.Get(ctx, p.options)
-	if err != nil {
-		return nil, graph.Stack(ctx, err)
-	}
-
-	return EmptyDeltaLinker[models.Eventable]{PageLinkValuer: resp}, nil
-}
-
-func (p *eventPager) setNext(nextLink string) {
-	p.builder = users.NewItemCalendarsItemEventsRequestBuilder(nextLink, p.gs.Adapter())
-}
-
-// non delta pagers don't need reset
-func (p *eventPager) reset(context.Context) {}
-
-func (p *eventPager) valuesIn(pl PageLinker) ([]getIDAndAddtler, error) {
-	return toValues[models.Eventable](pl)
-}
-
-// ---------------------------------------------------------------------------
-// delta item pager
-// ---------------------------------------------------------------------------
-
-var _ itemPager = &eventDeltaPager{}
-
-type eventDeltaPager struct {
-	gs          graph.Servicer
-	userID      string
-	containerID string
-	builder     *users.ItemCalendarsItemEventsDeltaRequestBuilder
-	options     *users.ItemCalendarsItemEventsDeltaRequestBuilderGetRequestConfiguration
-}
-
-func (c Events) NewEventDeltaPager(
-	ctx context.Context,
-	userID, containerID, oldDelta string,
-	immutableIDs bool,
-) (itemPager, error) {
-	options := &users.ItemCalendarsItemEventsDeltaRequestBuilderGetRequestConfiguration{
-		Headers: newPreferHeaders(preferPageSize(maxDeltaPageSize), preferImmutableIDs(immutableIDs)),
-	}
-
-	var builder *users.ItemCalendarsItemEventsDeltaRequestBuilder
-
-	if oldDelta == "" {
-		builder = getEventDeltaBuilder(ctx, c.Stable, userID, containerID, options)
-	} else {
-		builder = users.NewItemCalendarsItemEventsDeltaRequestBuilder(oldDelta, c.Stable.Adapter())
-	}
-
-	return &eventDeltaPager{c.Stable, userID, containerID, builder, options}, nil
-}
-
-func getEventDeltaBuilder(
-	ctx context.Context,
-	gs graph.Servicer,
-	userID, containerID string,
-	options *users.ItemCalendarsItemEventsDeltaRequestBuilderGetRequestConfiguration,
-) *users.ItemCalendarsItemEventsDeltaRequestBuilder {
-	// Graph SDK only supports delta queries against events on the beta version, so we're
-	// manufacturing use of the beta version url to make the call instead.
-	// See: https://learn.microsoft.com/ko-kr/graph/api/event-delta?view=graph-rest-beta&tabs=http
-	// Note that the delta item body is skeletal compared to the actual event struct.  Lucky
-	// for us, we only need the item ID.  As a result, even though we hacked the version, the
-	// response body parses properly into the v1.0 structs and complies with our wanted interfaces.
-	// Likewise, the NextLink and DeltaLink odata tags carry our hack forward, so the rest of the code
-	// works as intended (until, at least, we want to _not_ call the beta anymore).
-	rawURL := fmt.Sprintf(eventBetaDeltaURLTemplate, userID, containerID)
-	builder := users.NewItemCalendarsItemEventsDeltaRequestBuilder(rawURL, gs.Adapter())
-
-	return builder
-}
-
-func (p *eventDeltaPager) getPage(ctx context.Context) (DeltaPageLinker, error) {
-	resp, err := p.builder.Get(ctx, p.options)
-	if err != nil {
-		return nil, graph.Stack(ctx, err)
-	}
-
-	return resp, nil
-}
-
-func (p *eventDeltaPager) setNext(nextLink string) {
-	p.builder = users.NewItemCalendarsItemEventsDeltaRequestBuilder(nextLink, p.gs.Adapter())
-}
-
-func (p *eventDeltaPager) reset(ctx context.Context) {
-	p.builder = getEventDeltaBuilder(ctx, p.gs, p.userID, p.containerID, p.options)
-}
-
-func (p *eventDeltaPager) valuesIn(pl PageLinker) ([]getIDAndAddtler, error) {
-	return toValues[models.Eventable](pl)
-}
-
-func (c Events) GetAddedAndRemovedItemIDs(
-	ctx context.Context,
-	userID, containerID, oldDelta string,
-	immutableIDs bool,
-	canMakeDeltaQueries bool,
-) ([]string, []string, DeltaUpdate, error) {
-	ctx = clues.Add(ctx, "container_id", containerID)
-
-	pager, err := c.NewEventPager(ctx, userID, containerID, immutableIDs)
-	if err != nil {
-		return nil, nil, DeltaUpdate{}, graph.Wrap(ctx, err, "creating non-delta pager")
-	}
-
-	deltaPager, err := c.NewEventDeltaPager(ctx, userID, containerID, oldDelta, immutableIDs)
-	if err != nil {
-		return nil, nil, DeltaUpdate{}, graph.Wrap(ctx, err, "creating delta pager")
-	}
-
-	return getAddedAndRemovedItemIDs(ctx, c.Stable, pager, deltaPager, oldDelta, canMakeDeltaQueries)
 }
 
 // ---------------------------------------------------------------------------
