+++ src/pkg/services/m365/api/mail.go
@@ -17,7 +17,6 @@ import (
 	"github.com/alcionai/corso/src/pkg/backup/details"
 	"github.com/alcionai/corso/src/pkg/fault"
 	"github.com/alcionai/corso/src/pkg/logger"
-	"github.com/alcionai/corso/src/pkg/selectors"
 )
 
 const (
@@ -186,109 +185,6 @@ func (c Mail) PatchFolder(
 	}
 
 	return nil
-}
-
-// ---------------------------------------------------------------------------
-// container pager
-// ---------------------------------------------------------------------------
-
-type mailFolderPager struct {
-	service graph.Servicer
-	builder *users.ItemMailFoldersRequestBuilder
-}
-
-func (c Mail) NewMailFolderPager(userID string) mailFolderPager {
-	// v1.0 non delta /mailFolders endpoint does not return any of the nested folders
-	rawURL := fmt.Sprintf(mailFoldersBetaURLTemplate, userID)
-	builder := users.NewItemMailFoldersRequestBuilder(rawURL, c.Stable.Adapter())
-
-	return mailFolderPager{c.Stable, builder}
-}
-
-func (p *mailFolderPager) getPage(ctx context.Context) (PageLinker, error) {
-	page, err := p.builder.Get(ctx, nil)
-	if err != nil {
-		return nil, graph.Stack(ctx, err)
-	}
-
-	return page, nil
-}
-
-func (p *mailFolderPager) setNext(nextLink string) {
-	p.builder = users.NewItemMailFoldersRequestBuilder(nextLink, p.service.Adapter())
-}
-
-func (p *mailFolderPager) valuesIn(pl PageLinker) ([]models.MailFolderable, error) {
-	// Ideally this should be `users.ItemMailFoldersResponseable`, but
-	// that is not a thing as stable returns different result
-	page, ok := pl.(models.MailFolderCollectionResponseable)
-	if !ok {
-		return nil, clues.New("converting to ItemMailFoldersResponseable")
-	}
-
-	return page.GetValue(), nil
-}
-
-// EnumerateContainers iterates through all of the users current
-// mail folders, converting each to a graph.CacheFolder, and calling
-// fn(cf) on each one.
-// Folder hierarchy is represented in its current state, and does
-// not contain historical data.
-func (c Mail) EnumerateContainers(
-	ctx context.Context,
-	userID, baseContainerID string,
-	fn func(graph.CachedContainer) error,
-	errs *fault.Bus,
-) error {
-	el := errs.Local()
-	pgr := c.NewMailFolderPager(userID)
-
-	for {
-		if el.Failure() != nil {
-			break
-		}
-
-		page, err := pgr.getPage(ctx)
-		if err != nil {
-			return graph.Stack(ctx, err)
-		}
-
-		resp, err := pgr.valuesIn(page)
-		if err != nil {
-			return graph.Stack(ctx, err)
-		}
-
-		for _, fold := range resp {
-			if el.Failure() != nil {
-				break
-			}
-
-			if err := graph.CheckIDNameAndParentFolderID(fold); err != nil {
-				errs.AddRecoverable(ctx, graph.Stack(ctx, err).Label(fault.LabelForceNoBackupCreation))
-				continue
-			}
-
-			fctx := clues.Add(
-				ctx,
-				"container_id", ptr.Val(fold.GetId()),
-				"container_name", ptr.Val(fold.GetDisplayName()))
-
-			temp := graph.NewCacheFolder(fold, nil, nil)
-			if err := fn(&temp); err != nil {
-				errs.AddRecoverable(ctx, graph.Stack(fctx, err).Label(fault.LabelForceNoBackupCreation))
-				continue
-			}
-		}
-
-		link, ok := ptr.ValOK(page.GetOdataNextLink())
-		if !ok {
-			break
-		}
-
-		pgr.setNext(link)
-	}
-
-	return el.Failure()
 }
 
 // ---------------------------------------------------------------------------
@@ -547,161 +443,6 @@ func (c Mail) PostLargeAttachment(
 	}
 
 	return us, nil
-}
-
-// ---------------------------------------------------------------------------
-// item pager
-// ---------------------------------------------------------------------------
-
-var _ itemPager = &mailPager{}
-
-type mailPager struct {
-	gs      graph.Servicer
-	builder *users.ItemMailFoldersItemMessagesRequestBuilder
-	options *users.ItemMailFoldersItemMessagesRequestBuilderGetRequestConfiguration
-}
-
-func (c Mail) NewMailPager(
-	ctx context.Context,
-	userID, containerID string,
-	immutableIDs bool,
-) itemPager {
-	config := &users.ItemMailFoldersItemMessagesRequestBuilderGetRequestConfiguration{
-		QueryParameters: &users.ItemMailFoldersItemMessagesRequestBuilderGetQueryParameters{
-			Select: idAnd("isRead"),
-		},
-		Headers: newPreferHeaders(preferPageSize(maxNonDeltaPageSize), preferImmutableIDs(immutableIDs)),
-	}
-
-	builder := c.Stable.
-		Client().
-		Users().
-		ByUserId(userID).
-		MailFolders().
-		ByMailFolderId(containerID).
-		Messages()
-
-	return &mailPager{c.Stable, builder, config}
-}
-
-func (p *mailPager) getPage(ctx context.Context) (DeltaPageLinker, error) {
-	page, err := p.builder.Get(ctx, p.options)
-	if err != nil {
-		return nil, graph.Stack(ctx, err)
-	}
-
-	return EmptyDeltaLinker[models.Messageable]{PageLinkValuer: page}, nil
-}
-
-func (p *mailPager) setNext(nextLink string) {
-	p.builder = users.NewItemMailFoldersItemMessagesRequestBuilder(nextLink, p.gs.Adapter())
-}
-
-// non delta pagers don't have reset
-func (p *mailPager) reset(context.Context) {}
-
-func (p *mailPager) valuesIn(pl PageLinker) ([]getIDAndAddtler, error) {
-	return toValues[models.Messageable](pl)
-}
-
-// ---------------------------------------------------------------------------
-// delta item pager
-// ---------------------------------------------------------------------------
-
-var _ itemPager = &mailDeltaPager{}
-
-type mailDeltaPager struct {
-	gs          graph.Servicer
-	userID      string
-	containerID string
-	builder     *users.ItemMailFoldersItemMessagesDeltaRequestBuilder
-	options     *users.ItemMailFoldersItemMessagesDeltaRequestBuilderGetRequestConfiguration
-}
-
-func getMailDeltaBuilder(
-	ctx context.Context,
-	gs graph.Servicer,
-	user, containerID string,
-	options *users.ItemMailFoldersItemMessagesDeltaRequestBuilderGetRequestConfiguration,
-) *users.ItemMailFoldersItemMessagesDeltaRequestBuilder {
-	builder := gs.
-		Client().
-		Users().
-		ByUserId(user).
-		MailFolders().
-		ByMailFolderId(containerID).
-		Messages().
-		Delta()
-
-	return builder
-}
-
-func (c Mail) NewMailDeltaPager(
-	ctx context.Context,
-	userID, containerID, oldDelta string,
-	immutableIDs bool,
-) itemPager {
-	config := &users.ItemMailFoldersItemMessagesDeltaRequestBuilderGetRequestConfiguration{
-		QueryParameters: &users.ItemMailFoldersItemMessagesDeltaRequestBuilderGetQueryParameters{
-			Select: idAnd("isRead"),
-		},
-		Headers: newPreferHeaders(preferPageSize(maxDeltaPageSize), preferImmutableIDs(immutableIDs)),
-	}
-
-	var builder *users.ItemMailFoldersItemMessagesDeltaRequestBuilder
-
-	if len(oldDelta) > 0 {
-		builder = users.NewItemMailFoldersItemMessagesDeltaRequestBuilder(oldDelta, c.Stable.Adapter())
-	} else {
-		builder = getMailDeltaBuilder(ctx, c.Stable, userID, containerID, config)
-	}
-
-	return &mailDeltaPager{c.Stable, userID, containerID, builder, config}
-}
-
-func (p *mailDeltaPager) getPage(ctx context.Context) (DeltaPageLinker, error) {
-	page, err := p.builder.Get(ctx, p.options)
-	if err != nil {
-		return nil, graph.Stack(ctx, err)
-	}
-
-	return page, nil
-}
-
-func (p *mailDeltaPager) setNext(nextLink string) {
-	p.builder = users.NewItemMailFoldersItemMessagesDeltaRequestBuilder(nextLink, p.gs.Adapter())
-}
-
-func (p *mailDeltaPager) reset(ctx context.Context) {
-	p.builder = p.gs.
-		Client().
-		Users().
-		ByUserId(p.userID).
-		MailFolders().
-		ByMailFolderId(p.containerID).
-		Messages().
-		Delta()
-}
-
-func (p *mailDeltaPager) valuesIn(pl PageLinker) ([]getIDAndAddtler, error) {
-	return toValues[models.Messageable](pl)
-}
-
-func (c Mail) GetAddedAndRemovedItemIDs(
-	ctx context.Context,
-	userID, containerID, oldDelta string,
-	immutableIDs bool,
-	canMakeDeltaQueries bool,
-) ([]string, []string, DeltaUpdate, error) {
-	ctx = clues.Add(
-		ctx,
-		"category", selectors.ExchangeMail,
-		"container_id", containerID)
-
-	pager := c.NewMailPager(ctx, userID, containerID, immutableIDs)
-	deltaPager := c.NewMailDeltaPager(ctx, userID, containerID, oldDelta, immutableIDs)
-
-	return getAddedAndRemovedItemIDs(ctx, c.Stable, pager, deltaPager, oldDelta, canMakeDeltaQueries)
 }
 
 // ---------------------------------------------------------------------------
